Added some error checking to socket/* files
The socket/* files were from the alexa project and have a different way of handling error (namely exiting). May need to do a rewrite of socket/* to fit in with other style better.
However, DEBUGMSG for informational and ERROREXIT for full on errors is nicer. Other files could do with more DEBUGMSG informational points.